Substitute the Operand class by WireRef.
It make much more to have a fat Wire class which will store all
information about a wire.  For now, the information is just the unique
identifier, a name and the kind (quantum or classical), but later will
hold more, e.g., whether the wire is part of a register.

The previous Wire class is merged with Operand and renamed to WireRef.
This is much lighter representation which serve three purposes:
1) Uniquely identifying a wire within a circuit
2) Informing if a wire is a qubit or a cbit
3) When used withing controlled instructions, informing if the use of
   the wire is complemented or not, e.g., a CNOT can be negatively
   controlled by adding a NOT before and after the control.