Improvements for building wheels (#148)
This greatly improves the scripts for building wheels. Moreover, it expands
wheel building to support Apple Silicon, Linux POWER8 and Linux ARM64.
(Basically, most platforms supported by Qiskit.)

The s390x still a working in progress. I have not yet figure it out how
complicated it might be to support this platform, because I suspect some
code (either in tweedledum or in one of its dependencies) seems to
not be compatible with big endianness.

It also adds a workflow to run python tests.