Add support for Python 3.11 and drop 3.6 support
This commit updates the supported python versions in the trove
classifiers in the Python package metadata along with also updating the
test and wheel build matrix to support Python versions 3.7 through 3.11.
Python 3.6 which was previously supported has been dropped as it went
EoL in December 2021. Python 3.11 was recently released on 10-24-2022 so
it will be good to have precompiled binaries available for it.